+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Fixed
 
-- Address validation when there are null values.
+- Address validation when there are null values for optional fields.
 
 ## [4.12.0] - 2021-11-04